Molecule role testing added
For following roles:

- debug
- git
- include
- docker
- az cli
- google cloud
- docker-compose
- bash-it
- virtualbox
- terraform
- kubectl
- browsers
- misc - skips spotify due to use of snap package

added file check for git
added home dir creation for git

added linting for vs-code

lint playbook

added running_on_molecule to skip snap installs when running in containers

Added molecule  "role" which runs the entire playbook in the containers.
Could not find an easy way to set role path in molecule when importing playbook

Have not really used the "verify" stages as most of these roles are pacakge installs,
ansible checks the state of the packages as it goes.

Molecule tests on ubuntu 18/20 and mint 19.3 containers