WIP: Pipeline improvements
making init scripts more idempotent
init script is now broken down into functions, only loaded when
virtualenv is not detected.

Trying to kill sudo session before ansible run, as ansible is running
as root which messes up a lot of the config for zsh, and probably other
roles not yet found. This will need to be reviewed.